Split tests in separate file
